+++ examples/apps/lit-list/CHANGELOG.md
@@ -1,6 +1,6 @@
 # lit-todomvc
 
-## 0.1.5
+## 0.1.1
 
 ### Patch Changes
 
@@ -8,7 +8,7 @@
 - Updated dependencies
   - @inglorious/lit@1.3.0
 
-## 0.1.4
+## 0.1.1
 
 ### Patch Changes