+++ examples/apps/lit-router/CHANGELOG.md
@@ -1,5 +1,13 @@
 # lit-todomvc
 
+## 0.2.3
+
+### Patch Changes
+
+- Add list virtualization in @inglorious/lit
+- Updated dependencies
+  - @inglorious/lit@1.3.0
+
 ## 0.2.2
 
 ### Patch Changes